refactor: SessionManager가 Manger 인터페이스 구현하도록 변경
