+++ tomcat/src/main/java/nextstep/jwp/controller/ViewController.java
@@ -5,8 +5,8 @@ import static org.apache.coyote.http11.common.ContentType.HTML;
 import java.io.IOException;
 import java.util.Objects;
 import javassist.NotFoundException;
+import org.apache.catalina.SessionManager.Session;
 import org.apache.catalina.servlet.util.ResourceReader;
-import org.apache.coyote.http11.SessionManager.Session;
 import org.apache.coyote.http11.request.Request;
 import org.apache.coyote.http11.response.Response;
