+++ tomcat/src/main/java/org/apache/catalina/Manager.java
@@ -1,7 +1,8 @@
 package org.apache.catalina;
 
-import jakarta.servlet.http.HttpSession;
 import java.io.IOException;
+import java.util.Optional;
+import org.apache.catalina.SessionManager.Session;
 
 /**
  * A <b>Manager</b> manages the pool of Sessions that are associated with a particular Container.  Different Manager
@@ -26,7 +27,7 @@ public interface Manager {
      *
      * @param session Session to be added
      */
-    void add(HttpSession session);
+    void add(Session session);
 
     /**
      * Return the active Session, associated with this Manager, with the specified session id (if any); otherwise
@@ -38,12 +39,12 @@ public interface Manager {
      * @throws IllegalStateException if a new session cannot be instantiated for any reason
      * @throws IOException           if an input/output error occurs while processing this request
      */
-    HttpSession findSession(String id) throws IOException;
+    Optional<Session> findSession(String id) throws IOException;
 
     /**
      * Remove this Session from the active Sessions for this Manager.
      *
      * @param session Session to be removed
      */
-    void remove(HttpSession session);
+    void remove(Session session);
 }
